+++ CHANGELOG.md
@@ -15,5 +15,6 @@ and this project adheres to
 - ✨(users) make `email` our username
 - ✨(authentication) Support multiple identity providers
 - ✨(authentication) Encrypt identity providers' `extra_data`
+- ✨(OIDC Provider) Configure and customize the Authorization Server
 
 [unreleased]: https://github.com/suitenumerique/docs/compare/main